pyproject.toml
[tool.poetry]
name = "linkedin-researcher"
version = "0.1.0"
description = "Report generation based on LinkedIn profile."
authors = ["jaredkirby <jpeterkirby@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
